add logger , move worker number to worker service
