+++ workers-backend/src/worker/controllers/workers.controller.ts
@@ -16,19 +16,18 @@ import { TransformDataStructure } from '../../transformDataStructure/convertData
 import { Request, Response } from 'express';
 import { workerValidationsSchema } from '../validations/worker.validations.schema';
 import { ApiBody, ApiOperation } from '@nestjs/swagger';
+import { Logger } from '@nestjs/common';
 
 @Controller('workers')
 export class WorkersController {
+  private readonly logger = new Logger(WorkersController.name);
+
+  constructor(private readonly workersService: WorkersService) {}
+
   findAll(arg0: string) {
     throw new Error('Method not implemented.');
   }
-  constructor(private readonly workersService: WorkersService) {}
-
-  // @Get()
-  // async findAll(@Query('businessId') businessId: string): Promise<Employee[]> {
-  //   return this.workersService.findAllByBusinessId(businessId);
-  // }
-
+  
   @Get('employee/:id')
   getWorker(@Param('id') id: string) {
     return this.workersService.getEmployee(id);
@@ -60,32 +59,31 @@ export class WorkersController {
       },
     },
   })
+
   @Post('')
   async create(
     @Body(
       new ValidationPipe({
-        exceptionFactory: (errors) =>
-          new HttpException(errors, HttpStatus.BAD_REQUEST),
+        exceptionFactory: (errors) => {
+          Logger.log('error validation! '+errors);
+          return new HttpException(errors, HttpStatus.BAD_REQUEST);
+        },
       }),
     )
     requestBody: workerValidationsSchema,
   ): Promise<Employee> {
     try {
-      const workerCode = this.generateUniqueNumber();
-      requestBody.workerCode = workerCode;
       const result = await this.workersService.createEmployee(requestBody);
+      this.logger.log("good");
       return result;
     } catch (error) {
-      if (error instanceof HttpException) throw error;
+      if (error instanceof HttpException) {
+        throw error;
+      }
       throw new HttpException(
         'Internal server error',
         HttpStatus.INTERNAL_SERVER_ERROR,
       );
     }
-  }
-  generateUniqueNumber(): string {
-    const timestamp = new Date().getTime(); // Get current timestamp
-    const random = Math.floor(Math.random() * 10000); // Generate random number between 0 and 9999
-    return `${timestamp}${random}`; // Concatenate timestamp and random number
   }
 }
